### 2025-12-22 #9683 @editorialbot
[PRE REVIEW]: dcTMD: a python package for performing dissipation-corrected targeted molecular dynamics
**Submitting author:** @floWneffetS (<a href="http://orcid.org/0000-0003-1752-6175">Steffen Wolf</a>)
**Repository:** https://github.com/moldyn/dcTMD
**Branch with paper.md** (empty if default branch):
**Version:** v0.4.1
**Managing EiC:** Rachel Kurchin
