### 2026-02-27 712c206
update paper to JOSSs updated submission requirements and scope criteria
