+++ paper/paper.md
@@ -8,36 +8,59 @@ authors:
   - name: Daniel Nagel
     affiliation: 2
   - name: Steffen Wolf
-    affiliation: 1
+    affiliation: 1,3
 affiliations:
   - index: 1
     name: University of Freiburg, Institute of Physics, Freiburg, Germany
   - index: 2
     name: University of Heidelberg, Institute of Theoretical Physics, Heidelberg, Germany
-date: "2025-11-27"
+  - index: 3
+    name: Darmstadt University of Applied Science, Faculty of Chemical Engineering and Biotechnology, Darmstadt, Germany
+date: "2025-01-27"
 bibliography: paper.bib
 ---
 
 
 # Summary
-`dcTMD` is a Python package designed to extract free-energy and non-equilibrium friction estimates from targeted molecular dynamics (TMD) simulations [@Schlitter1994_TMD]. The method implemented here is called *dissipation-corrected targeted molecular dynamics* (dcTMD) by @Wolf2018_dcTMD.
-Given a set of non-equilibrium simulations where, for example, a ligand is pulled from a binding site via a velocity constraint, this tool performs automated post-processing of the bias-force time traces to estimate the underlying free-energy landscape and the friction (dissipation) along the unbinding coordinate.  
+`dcTMD` is a Python package designed to extract free-energy and non-equilibrium friction estimates from targeted molecular dynamics (TMD) simulations by @Schlitter1994_TMD. The method implemented here is called *dissipation-corrected targeted molecular dynamics* (dcTMD) by @Wolf2018_dcTMD.
+Given a set of non-equilibrium simulations where, for example, a ligand is pulled from a protein binding site via a velocity constraint, this tool performs automated post-processing of the bias-force time traces to estimate the underlying free-energy landscape and the friction (dissipation) along the unbinding coordinate.
 
-The method is based on a second-order cumulant expansion of *Jarzynski’s equality* [@Jarzynski1997_NonEq], which connects non-equilibrium work distributions to equilibrium free-energy differences. Combined with a Markovian Langevin Equation, dcTMD further allows the extraction of *position- and velocity-dependent friction coefficients* from the same non-equilibrium data. This approach has been succesfully applied in multiple studies (@Wolf2020_MultisecondDissociation @Jaeger2022_IonChannelConductance @Post2022_JCTC @Taenzel2024_CommunityPaths @Jaeger2025_SimilarityMeasures @Milster_2025_NEQ).
+The method is based on a second-order cumulant expansion of the *Jarzynski equality* (@Jarzynski1997_NonEq), which connects non-equilibrium work distributions to equilibrium free-energy differences. Combined with a Markovian Langevin Equation, dcTMD further allows the extraction of *position- and velocity-dependent friction coefficients* from the same non-equilibrium data. This approach has been succesfully applied in multiple studies (@Wolf2020_MultisecondDissociation, @Jaeger2022_IonChannelConductance, @Post2022_JCTC, @Taenzel2024_CommunityPaths, and @Jaeger2025_SimilarityMeasures @Milster_2025_NEQ).
 The resulting free-energy and friction profiles can subsequently be used to estimate *binding and unbinding rate constants* following @Wolf2020_MultisecondDissociation.  
 
 The software is intended for molecular dynamics practitioners interested in ligand–protein unbinding, mechanistic interpretation of binding kinetics, and quantitative modeling of non-equilibrium effects in soft condensed matter and biomolecular systems.  
 
 # Statement of need
[…]
