+++ paper/paper.bib
@@ -138,7 +138,7 @@
 	number = {5},
 	journaltitle = {J. Chem. Theory Comput.},
 	year = {2022},
-	doi = {10.1021/acs.jctc.3c01017}
+	doi = {10.1021/acs.jctc.2c00190}
 }
 
 @article{Bray_2022_JCIM,
@@ -221,10 +221,94 @@
 	issue        = {7825}
 }
 
+@article{abraham2015gromacs,
+	title        = {GROMACS: High performance molecular simulations through multi-level parallelism from laptops to supercomputers},
+	author       = {Mark J. Abraham and Teemu Murtola and Roland Schulz and Szilárd Páll and Jeremy C. Smith and Berk Hess and Erik Lindahl},
+	year         = {2015},
+	journal      = {SoftwareX},
+	volume       = {1--2},
+	pages        = {19--25},
+	doi          = {10.1016/j.softx.2015.06.001},
+	issn         = {2352-7110}
+}
+
 @software{click,
   author = {{The Pallets Projects}},
   title = {{Click: Python composable command line interface toolkit}},
   url = {https://palletsprojects.com/p/click},
   version = {8.1.*},
   year = {2023}
+}
+
+@article{zubieta2024pysages,
+  title={PySAGES: flexible, advanced sampling methods accelerated with GPUs},
+  author={Zubieta Rico, Pablo F and Schneider, Ludwig and P{\'e}rez-Lemus, Gustavo R and Alessandri, Riccardo and Dasetty, Siva and Nguyen, Trung D and Men{\'e}ndez, Cintia A and Wu, Yiheng and Jin, Yezhi and Xu, Yinan and others},
+  journal={npj Computational Materials},
+  volume={10},
+  number={1},
+  pages={35},
+  year={2024},
+  publisher={Nature Publishing Group UK London}
+}
+
+
+@article{fiorin2013using,
+  title={Using collective variables to drive molecular dynamics simulations},
+  author={Fiorin, Giacomo and Klein, Michael L and H{\'e}nin, J{\'e}r{\^o}me},
+  journal={Molecular Physics},
+  volume={111},
+  number={22-23},
+  pages={3345--3362},
+  year={2013},
+  publisher={Taylor \& Francis}
+}
+
+@article{fiorin2024expanded,
+  title={Expanded functionality and portability for the Colvars library},
+  author={Fiorin, Giacomo and Marinelli, Fabrizio and Forrest, Lucy R and Chen, Haochuan and Chipot, Christophe and Kohlmeyer, Axel and Santuz, Hubert and H{\'e}nin, J{\'e}r{\^o}me},
+  journal={The Journal of Physical Chemistry B},
+  volume={128},
+  number={45},
+  pages={11108--11123},
+  year={2024},
+  publisher={ACS Publications},
+  doi={ 10.1021/acs.jpcb.4c05604},
+}
+
+@article{plumed2019promoting,
+  title   = {Promoting transparency and reproducibility in enhanced molecular simulations},
+  author  = {{The PLUMED consortium}},
+  journal = {Nature Methods},
+  year    = {2019},
+  volume  = {16},
+  number  = {8},
+  pages   = {670--673},
+  doi={10.1038/s41592-019-0506-8},
+}
+
+@article{scikit-learn,
+title={Scikit-learn: Machine Learning in {P}ython},
+author={Pedregosa, F. and Varoquaux, G. and Gramfort, A. and Michel, V.
+and Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P.
+and Weiss, R. and Dubourg, V. and Vanderplas, J. and Passos, A. and
[…]
