### 2026-02-27 bdf590d
Update paper.md
