### 2026-02-27 58c49e7
Update paper.md
