### 2026-02-27 #9683 @floWneffetS
> [@floWneffetS](https://github.com/floWneffetS) please let us know when you've updated your paper.

Dear @arfon , please excuse the delay - our manuscript has now been updated and changed according to JOSS's new format changes. Please go ahead with the reviewing process.
