### 2026-03-22 #9683 @abhishektiwari
👋 @j-wags @jkrajniak @colinbousige, would any of you be willing to review this submission for Journal of Open Source Software (JOSS)? The submission under consideration is "dcTMD: a python package for performing dissipation-corrected targeted molecular dynamics".

Thanks
Abhi
