### 2026-03-22 #9683 @jkrajniak
> 👋 @j-wags @jkrajniak @colinbousige, would any of you be willing to review this submission for Journal of Open Source Software (JOSS)? The submission under consideration is "dcTMD: a python package for performing dissipation-corrected targeted molecular dynamics".
>
> We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html.
>
> Thanks
> Abhi

yes, add me
