### 2026-03-23 #9683 @chrisjonesBSU
> 👋 [@chrisjonesBSU](https://github.com/chrisjonesBSU) [@hejung](https://github.com/hejung) [@AnjaConev](https://github.com/AnjaConev), would any of you be willing to review this submission for Journal of Open Source Software (JOSS)? The submission under consideration is "dcTMD: a python package for performing dissipation-corrected targeted molecular dynamics".
>
> We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html).
>
> Thanks Abhi

Hello @abhishektiwari, I can be a reviewer for this, please add me.
