### 2026-03-23 #10244 @editorialbot
[REVIEW]: dcTMD: a python package for performing dissipation-corrected targeted molecular dynamics
**Submitting author:** @floWneffetS (<a href="http://orcid.org/0000-0003-1752-6175">Steffen Wolf</a>)
**Repository:** https://github.com/moldyn/dcTMD
**Branch with paper.md** (empty if default branch):
**Version:** v0.4.1
