### 2026-04-24 #10244 @chrisjonesBSU
Hello  @abhishektiwari and @floWneffetS. I just wanted to chime in and apologize about the delay. I've had some urgent priorities come up and some holiday travel over the last month. I'm just about done working my way through reviewing the dcTMD repository and documentation and will then shift focus to the paper.  I'll plan on having my initial review submitted sometime next week.
