### 2026-06-08 158841c
add further type hints.
