+++ src/dcTMD/dcTMD.py
@@ -19,6 +19,7 @@ from beartype.typing import Union, Tuple, Optional
 from sklearn.base import BaseEstimator, TransformerMixin
 
 from dcTMD import utils
+from dcTMD.storing import WorkSet, ForceSet
 from dcTMD._typing import (
     Int,
     Float,
@@ -32,8 +33,8 @@ from dcTMD._typing import (
 
 
 class _SmoothBasisEstimator(ABC):
-    """Class with the smoothing method for `WorkEstimator`, `ForceEstimator`.
-    """
+    """Class with the smoothing method for `WorkEstimator`, `ForceEstimator`."""
+
     @beartype
     def smooth_friction(
         self,
@@ -73,7 +74,9 @@ class _SmoothBasisEstimator(ABC):
 
 
 class WorkEstimator(
-    _SmoothBasisEstimator, TransformerMixin, BaseEstimator,
+    _SmoothBasisEstimator,
+    TransformerMixin,
+    BaseEstimator,
 ):  # noqa: WPS230, WPS214
     """Class for performing dcTMD analysis on a work set.
 
@@ -151,7 +154,7 @@ dcTMD.WorkEstimator.estimate_free_energy_errors].
     @beartype
     def fit(
         self,
-        work_set,
+        work_set: WorkSet,
     ):
         """
         Estimate free energy and friction.
@@ -177,15 +180,17 @@ dcTMD.WorkEstimator.estimate_free_energy_errors].
 
     @beartype
     def transform(
-        self, X, y=None,
-    ) -> Tuple[Float1DArray, Float1DArray]:  # noqa: WPS111
+        self,
+        X: Optional[WorkSet] = None,
+        y: Optional[np.ndarray] = None,
+    ) -> Tuple[Float1DArray, Float1DArray]:
         """Return free energy and friction estimates."""
         return self.dG_, self.friction_
 
     @beartype
     def estimate_free_energy(
         self,
-        work_set=None,
+        work_set: Optional[Float2DArray] = None,
     ) -> Tuple[Float1DArray, Float1DArray, Float1DArray]:
         """
         Estimate free energy.
@@ -193,8 +198,9 @@ dcTMD.WorkEstimator.estimate_free_energy_errors].
         Parameters
         ----------
         work_set : optional
-            Instance of a WorkSet containing constraint forces, for which the
-            free energy and friction are estimated.
+            2D array of work time traces, for which the free energy and
+            friction are estimated. Passed when called during bootstrapping;
+            otherwise the work traces of the fitted WorkSet are used.
 
         Returns
         -------
@@ -216,6 +222,7 @@ dcTMD.WorkEstimator.estimate_free_energy_errors].
             is_bootstrapping = True
 
         from scipy.constants import R  # noqa: WPS347
+
         RT = R * self.temperature / 1e3
 
         W_mean = np.mean(work_set, axis=0)
@@ -235,10 +242,11 @@ dcTMD.WorkEstimator.estimate_free_energy_errors].
         n_resamples: Int,
         mode: Union[StrStd, NumInRange0to1],
         seed: Optional[Int] = None,
-    ) -> Tuple[Union[Float1DArray, Float2DArray],
-               Union[Float1DArray, Float2DArray],  # noqa: WPS318
-               Union[Float1DArray, Float2DArray],
-               ]:
+    ) -> Tuple[
+        Union[Float1DArray, Float2DArray],
+        Union[Float1DArray, Float2DArray],  # noqa: WPS318
+        Union[Float1DArray, Float2DArray],
+    ]:
         """
         Estimate bootstrapping errors for the free energy estimate.
 
@@ -288,9 +296,11 @@ estimate_free_energy].
         self,
     ) -> None:
[…]
