+++ .github/workflows/pytest.yml
@@ -29,7 +29,7 @@ jobs:
       run: |
         pytest --cov-report=xml
     - name: Upload coverage to Codecov
-      uses: codecov/codecov-action@v6
+      uses: codecov/codecov-action@v7
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
         files: ./coverage.xml
