### 2026-07-10 bd99018
Reference/DOI fixes
