+++ paper/paper.bib
@@ -130,17 +130,6 @@
 	year = {2023}
 }
 
-@article{Post_2022_JCTC,
-	title = {Molecular origin of driving-dependent friction in fluids},
-	author = {Post, Matthias and Wolf, Steffen and Stock, Gerhard},
-	volume = {18},
-	pages = {2816--2825},
-	number = {5},
-	journaltitle = {J. Chem. Theory Comput.},
-	year = {2022},
-	doi = {10.1021/acs.jctc.2c00190}
-}
-
 @article{Bray_2022_JCIM,
 	title = {Ligand Unbinding Pathway and Mechanism Analysis Assisted by Machine Learning and Graph Methods},
 	author = {Bray, Simon and Tänzel, Victor and Wolf, Steffen},
@@ -248,6 +237,7 @@
   number={1},
   pages={35},
   year={2024},
+  doi = {10.1038/s41524-023-01189-z},
   publisher={Nature Publishing Group UK London}
 }
 
@@ -260,6 +250,7 @@
   number={22-23},
   pages={3345--3362},
   year={2013},
+  doi = {10.1080/00268976.2013.813594},
   publisher={Taylor \& Francis}
 }
 
@@ -296,7 +287,7 @@ journal={Journal of Machine Learning Research},
 volume={12},
 pages={2825--2830},
 year={2011},
-doi={10.5555/1953048.2078195},
+url={https://jmlr.org/papers/v12/pedregosa11a.html},
 }
 
 @article{Lalis2025,
