### 2026-07-10 9610b4d
Document path resolution for CLI list files
