### 2026-07-10 4d303ab
Add citation and info about trypsin benzamidine simulation and a system render.
