### 2026-07-10 53640de
Add README.md with simulation details to testdata
