### 2026-07-10 a502fe5
Add explicit dependency files for pip and conda.
