### 2026-07-10 459b031
Include dependency files in source distributions  via MANIFEST.in
