### 2026-08-07 884b947
Update Gromacs.md
added link to the NOMAD pulling force file repository.
