### 2026-08-18 8f51fb9
Update paper.md
