+++ paper/paper.md
@@ -53,7 +53,7 @@ Finally, the `featureset` enables pathway-level analysis by loading trajectory-s
 
 
 # Research impact statement
-As mentioned above, dcTMD has been already shown to efficiently coarse-grain system dynamics for such diverse systems as protein-ligand complexes (@Wolf2020_MultisecondDissociation), ion channels (@Jaeger2022_IonChannelConductance) or to gain insight into shear-dependent viscosities in solvents and lubricants (@Post2022_JCTC, @Milster_2025_NEQ). The method thus has proven its capabilities to the scientific community. Its distribution so far was hampered by the necessity to manually implement dcTMD in the form of self-written scripts. The presented package represents a comprehensive and easy-to-use tool to allow other groups who are interested in applying dcTMD to their problems, that has been optimized for fast and efficient data evaluation with minimal runtime and memory requirements. We allow not only for dissipation correction or friction coefficient calculation via a work-based analysis, but to gain insight into bath degrees of freedom via a force-correlation analysis. Furthermore, the standardised outputs allow for data evaluation quality control across research groups and allow for the establishment of best practices in the usage of dcTMD. The software has already been successfully applied in the following studies: @Taenzel2024_CommunityPaths, @Lalis2025, @Jaeger2025_SimilarityMeasures, and @Milster_2025_NEQ. 
+As mentioned above, dcTMD has been already shown to efficiently coarse-grain system dynamics for such diverse systems as protein-ligand complexes (@Wolf2020_MultisecondDissociation), ion channels (@Jaeger2022_IonChannelConductance) or to gain insight into shear-dependent viscosities in solvents and lubricants (@Post2022_JCTC, @Milster_2025_NEQ). The method thus has proven its capabilities to the scientific community. Its distribution so far was hampered by the necessity to manually implement dcTMD in the form of self-written scripts. The presented package represents a comprehensive and easy-to-use tool to allow other groups who are interested in applying dcTMD to their problems, that has been optimized for fast and efficient data evaluation with minimal runtime and memory requirements. We allow not only for dissipation correction or friction coefficient calculation via a work-based analysis, but to gain insight into bath degrees of freedom, i.e., microscopic degrees of freedom that have been integrated out during calculation of the potential of mean force, via a force-correlation analysis. Furthermore, the standardised outputs allow for data evaluation quality control across research groups and allow for the establishment of best practices in the usage of dcTMD. The software has already been successfully applied in the following studies: @Taenzel2024_CommunityPaths, @Lalis2025, @Jaeger2025_SimilarityMeasures, and @Milster_2025_NEQ. 
 
 # Implementation
 The code is written in Python (versions 3.9–3.14) and is available under the MIT license.  
