### 2026-08-18 #10244 @floWneffetS
> [@abhishektiwari](https://github.com/abhishektiwari) yes, I completed my review and checked every item in the list. I can recommend for acceptance! Although, there appears to be 2 open PRs ([moldyn/dcTMD#43](https://github.com/moldyn/dcTMD/pull/43), [moldyn/dcTMD#42](https://github.com/moldyn/dcTMD/pull/42)) related to these reviews, so hopefully they can be merged sooner rather than later.
>
> Also, the typo I mentioned in one of my previous comments is still present in the paper.md file:
>
> > I've read through the paper and don't have any suggested major changes. I did find one typo in this sentence starting on line 84 (typo is on line 86 of the pdf)
>
> > > We allow not only for dissipation correction or friction coefficient calculation via a work-based analysis, but to gain insight into bath degrees of freedom via a force-correlation analysis
>
> > bath should be changed to both.

Dear @chrisjonesBSU ,
please excuse the delay – I now checked the typo you mentioned, but it turns out that this actually not a typo but correct. The "bath" here refers to a "system/bath" separation, and thus microscopic degrees of freedom that have been integrated out. I have improved the explanation by changing the respective sentence into: "We allow not only for dissipation correction or friction coefficient calculation via a work-based analysis, but to gain insight into bath degrees of freedom, i.e., microscopic degrees of freedom that have been integrated out during calculation of the potential of mean force, via a force-correlation analysis."
Furthermore, the two pull requests have been checked and merged now.
