### 2026-08-24 #10244 @floWneffetS
> > [@hejung](https://github.com/hejung) - We are now 5 months mark for this review - appreciate if you can confirm if you are still able to progress with the review in next few day. I have not heard from you since June. Hope all is well. Otherwise I would like to move ahead with 2 completed reviews. Thanks for your support.
>
> Very sorry for the delay, I just came back from summer vacations. I am happy to have a final look tomorrow, but if you want progress with the two reviews already. As far as I remember the only somewhat blocking (according to the review guidelines) was that I would not be able to reproduce the figures from the publication. And as nice-to-haves (not blocking but strongly recommended), I remember that it would be nice to make clear from which molecular systems the pulling traces in the test files and example notebooks are.

Dear @hejung , concerning the nice-to-haves: we now include the information "This tutorial allows targeted MD simulations of a trypsin-benzamidine complex." at the beginning of the Gromacs tutorial in https://moldyn.github.io/dcTMD/tutorials/Gromacs/. Furthermore, at the end of the same page, we mention that: "If you are interested in performing an example analysis, but do not have the HPC capacities to perform a sufficient number of simulations, you can find an example set of 84 trypsin-benzamidine unbinding simulation force files generated from the above tutorial at [this NOMAD repository](https://nomad-lab.eu/prod/v1/gui/upload/id/BG7TgBEsR46vYAEiHDv6Hg)."
