### 2026-08-30 #10244 @hejung
I just finished my review. Thank you all for waiting for me and sorry for the delay!

Especially with the now included NOMAD repository, everything looks great and I am recommending to accept the paper for publication.
