+++ paper/paper.md
@@ -97,7 +97,7 @@ Each trajectory contains the constraint force $f(t)$ from which the work along t
 `dcTMD` provides plotting tools for work distribution analysis, free-energy $\Delta G(x)$ and friction profiles $\Gamma(x)$.
 
 #### 4. Example
-![Figures crated using data taken from @Wolf2020_MultisecondDissociation of trypsin-benzamidine unbinding.  a)-c) work distribution analysis. d) Decomposition of mean work $W_{\rm mean} = \langle W(x) \rangle$ into free energy $\Delta G(x)$ and dissipation work $W_{\rm diss}(x)$. e) non-equilibrium friction coefficient $\Gamma (x)$ along the pulling coordinate $x$.](figures/image2.png){width=\linewidth}
+![Figures created using data taken from @Wolf2020_MultisecondDissociation of trypsin-benzamidine unbinding.  a)-c) work distribution analysis. d) Decomposition of mean work $W_{\rm mean} = \langle W(x) \rangle$ into free energy $\Delta G(x)$ and dissipation work $W_{\rm diss}(x)$. e) non-equilibrium friction coefficient $\Gamma (x)$ along the pulling coordinate $x$.](figures/image2.png){width=\linewidth}
 
 Figure 1 displays a common analysis of a set of unbinding trajectories from TMD simulations of the trypsin-benzamidine complex [@Wolf2020_MultisecondDissociation]. The analysis of the work distribution displays good agreement with a normal distribution at two different evaluated positions of the pulling coordinate $x$. The mean work $W_{\rm mean} = \langle W(x) \rangle$, which shows no features on its own, yields a free energy profile $\Delta G(x)$, which displays a clearly defined transition state at $x \approx 0.45$ nm as well as a bound state in form of a free energy minimum at $x \approx 0.0$ nm and an unbound continuum for $x > 0.6$ nm. The maximum in friction $\Gamma$ around $x = 0.5$ nm is indicative of changes in the hydration of both ligand and binding site.
 
