### 2026-09-07 8654311
Update paper.md
added ORCID IDs.
