### 2026-09-07 e4922a4
bump v1.0.0
