+++ setup.py
@@ -44,7 +44,7 @@ README = (HERE / 'README.md').read_text()
 
 setup(
     name='dcTMD',
-    version='0.4.1',
+    version='1.0.0',
     description='Analyse targeted molecular dynamics data with dcTMD',
     long_description=README,
     long_description_content_type='text/markdown',
