+++ src/dcTMD/__init__.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
-""" --8<-- "docs/dcTMD.md" """
+"""--8<-- "docs/dcTMD.md" """
 
-__all__ = ['dcTMD', 'storing', 'io', 'utils', 'featureset']
+__all__ = ["dcTMD", "storing", "io", "utils", "featureset"]
 
 
 from dcTMD import dcTMD, io, storing, utils, featureset
 
-__version__ = '0.4.1'
+__version__ = "1.0.0"
